+++ CHANGELOG.md
@@ -47,6 +47,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Command line option to export causative variants by category and/or genome build (#6202)
 - An additional button on causatives and verified pages to download SNVs and SVs as input for the managed variants list (admins only) (#6205 and #6231)
 - Display total number of variants and number of variants returned by a filter on Managed Variants page (#6223)
+- A `--liftover-from` option to the `export managed` command line, so managed variants can be exported and imported again in another genome build (#6225)
 ### Changed
 - Genome build is now shown on variant verification "Sanger" emails (#6194)
 - Refactor, speedup (dry-run only) and add a progress bar to `scout delete variants` cmd (#6094)